moved lint in pkg and added tests
