+++ cmd/staticlint/main.go
@@ -1,8 +1,7 @@
 package main
 
 import (
-	"go/ast"
-
+	"github.com/UndeadDemidov/yandex-praktikum/pkg/linter"
 	"github.com/fatih/errwrap/errwrap"
 	"github.com/masibw/goone"
 	"golang.org/x/tools/go/analysis"
@@ -46,7 +45,6 @@ import (
 	"golang.org/x/tools/go/analysis/passes/unsafeptr"
 	"golang.org/x/tools/go/analysis/passes/unusedresult"
 	"golang.org/x/tools/go/analysis/passes/unusedwrite"
-	"golang.org/x/tools/go/ast/inspector"
 	"honnef.co/go/tools/staticcheck"
 	"honnef.co/go/tools/stylecheck"
 )
@@ -64,68 +62,8 @@ func main() {
 	mychecks = appendPasses(mychecks)
 	mychecks = append(mychecks, errwrap.Analyzer)
 	mychecks = append(mychecks, goone.Analyzer)
-	mychecks = append(mychecks, ExitInMainAnalyzer)
+	mychecks = append(mychecks, linter.ExitInMainAnalyzer)
 	multichecker.Main(mychecks...)
-}
-
-var ExitInMainAnalyzer = &analysis.Analyzer{
-	Name:     "exitinmain",
-	Doc:      "check for os.Exit call in func main in package main",
-	Requires: []*analysis.Analyzer{inspect.Analyzer},
-	Run:      run,
-}
-
-func run(pass *analysis.Pass) (interface{}, error) {
-	isMainPkg := func(x *ast.File) bool {
-		return x.Name.Name == "main"
-	}
-
-	isMainFunc := func(x *ast.FuncDecl) bool {
-		return x.Name.Name == "main"
-	}
-
-	isOsExit := func(x *ast.SelectorExpr, isMain bool) bool {
-		if !isMain || x.X == nil {
-			return false
-		}
-		ident, ok := x.X.(*ast.Ident)
-		if !ok {
-			return false
-		}
-		if ident.Name == "os" && x.Sel.Name == "Exit" {
-			pass.Reportf(ident.NamePos, "os.Exit called in main func in main package")
-			return true
-		}
-		return false
-	}
-
-	i := pass.ResultOf[inspect.Analyzer].(*inspector.Inspector)
-	nodeFilter := []ast.Node{
-		(*ast.File)(nil),
-		(*ast.FuncDecl)(nil),
-		(*ast.SelectorExpr)(nil),
-	}
-	mainInspecting := false
-	i.Preorder(nodeFilter, func(n ast.Node) {
-		switch x := n.(type) {
-		case *ast.File:
-			if !isMainPkg(x) { // если пакет на main - выходим
-				return
-			}
-		case *ast.FuncDecl: // определение функции
-			f := isMainFunc(x)
-			if mainInspecting && !f { // если до этого инспектировали main, а теперь нет - можно заканчивать
-				return
-			}
-			mainInspecting = f
-		case *ast.SelectorExpr:
-			if isOsExit(x, mainInspecting) {
-				return
-			}
-		}
-	})
-
-	return nil, nil
 }
 
 func appendPasses(in []*analysis.Analyzer) []*analysis.Analyzer {
