+++ pkg/linter/osexit_test.go
@@ -1,0 +1,11 @@
+package linter
+
+import (
+	"testing"
+
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func Test_ExitInMainAnalyzer(t *testing.T) {
+	analysistest.Run(t, analysistest.TestData(), ExitInMainAnalyzer, "./...")
+}
