+++ pkg/linter/testdata/main/osexitinmain.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"os"
+)
+
+func main() {
+	println("here is it")
+	os.Exit(1) // want `os.Exit called in main func in main package`
+}
